+++ src/commands.ts
@@ -6,7 +6,7 @@ import { createLogger } from './logger'
 import Mru from './model/mru'
 import { toArray } from './util/array'
 import { Extensions as ExtensionsInfo, IExtensionRegistry } from './util/extensionRegistry'
-import { getExtensionId } from './util/extensionId'
+import { getExtensionId, prefixExtensionError } from './util/extensionId'
 import { Disposable } from './util/protocol'
 import { Registry } from './util/registry'
 import { toText } from './util/string'
@@ -44,10 +44,7 @@ class CommandItem implements Disposable, Command {
 
   private withExtension(e: any): any {
     let extensionId = getExtensionId(this.impl) || getExtensionId(this.thisArg)
-    if (extensionId && e instanceof Error && !e.message.startsWith('[extension:')) {
-      e.message = `[extension: ${extensionId}] ${e.message}`
-    }
-    return e
+    return extensionId ? prefixExtensionError(e, extensionId) : e
   }
 
   public dispose(): void {